requirements.txt
Flask
Flask-SocketIO
eventlet
pymongo[srv]
dnspython
python-dotenv
gunicorn
